requirements.txt
streamlit
openai
sqlalchemy
langchain
PyPDF2
python-docx
python-dotenv
